+++ CHANGELOG.md
@@ -10,8 +10,11 @@ GAMSPy 0.11.0
   - Find variables in equations by iteratively traversing instead of doing recursion.
 - Documentation
   - Add documentation of GAMS MIRO integration.
+  - Add documentation for NEOS backend.
 - Testing
   - Add tests for GAMS MIRO.
+  - Add NEOS Server as a backend to solve models.
+  - Add tests for NEOS backend.
 
 -------------------------------------------------------------------------------
 GAMSPy 0.10.5